Remove "(CLI)" from aws module doc
Turns out the parentheses were interpreted by antsy as being an
ANSI control string (albeit invalid) so an error was raised.
We *could* fix the ok modules info doc parser to use different
delimeters, but since it's inessential to have the acronym there
we'll just take it out for now.